Updating algorithm for next token
