+++ README.md
@@ -1,3 +1,9 @@
 # PythonDOMParser
 
+## Notes
++ For HTML5
+    + If the ```<!DOCTYPE html>``` goes to standard mode
+    + If not, goes into quirks mode
+    + These modes are the same in this
+
 ## Limitations
