+++ src/html/index.html
@@ -1,4 +1,10 @@
-<html id="html" class="html-class"> 
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title>Document</title>
+    </head>
     <body id="body">
         <div class="div">
             Div 1
