+++ src/lexer/stream_reader.py
