+++ src/tables/classifier_table.csv
@@ -72,3 +72,6 @@ _,underscore
 =,equal
 ',sinlge_quotation_mark
 \",double_quotation_mark
+!,exclamation_mark
+.,dot
+",",comma
