+++ src/tables/token_type_table.csv
@@ -1,4 +1,5 @@
 State,Type
-OPEN_START_TAG_FROM_ATTRIBUTE,ATTRIBUTE
-CLOSE_START_TAG,OPENING_TAG
-CLOSE_END_TAG,CLOSING_TAG
+OPEN_START_TAG,OPEN_START_TAG
+CLOSE_TAG,CLOSE_TAG
+STRING,STRING
+EQUAL,EQUAL
