--- src/tables/token_type_table_new.csv
@@ -1,6 +1,0 @@
-State,Type
-OPEN_TAG,OPEN_TAG
-OPEN_END_TAG,OPEN_END_TAG
-CLOSE_TAG,CLOSE_TAG
-END_STRING,END_STRING
-EQUAL,EQUAL
