+++ src/tables/transition_table.csv
@@ -1,31 +1,31 @@
 Start_State,On,End_State
-START,open_brace,OPEN_START_TAG
-OPEN_START_TAG,letter,OPEN_START_TAG_NAME
-OPEN_START_TAG_NAME,letter,OPEN_START_TAG_NAME
-OPEN_START_TAG_NAME,digit,OPEN_START_TAG_NAME
-OPEN_START_TAG_NAME,underscore,OPEN_START_TAG_NAME
-OPEN_START_TAG_NAME,hyphen,OPEN_START_TAG_NAME
-OPEN_START_TAG_NAME,space,OPEN_START_TAG_NAME
-OPEN_START_TAG_NAME,close_brace,CLOSE_START_TAG
-OPEN_START_TAG_NAME,forward_slash,OPEN_END_TAG
-OPEN_START_TAG,forward_slash,OPEN_END_TAG
-OPEN_END_TAG,letter,OPEN_END_TAG_NAME
-OPEN_END_TAG_NAME,letter,OPEN_END_TAG_NAME
-OPEN_END_TAG_NAME,digit,OPEN_END_TAG_NAME
-OPEN_END_TAG_NAME,underscore,OPEN_END_TAG_NAME
-OPEN_END_TAG_NAME,hyphen,OPEN_END_TAG_NAME
-OPEN_END_TAG_NAME,space,OPEN_END_TAG_NAME
-OPEN_END_TAG_NAME,close_brace,CLOSE_END_TAG
+
 START,space,START
-START,letter,TEXT
-START,digit,TEXT
-START,hyphen,TEXT
-START,underscore,TEXT
-TEXT,letter,TEXT
-TEXT,digit,TEXT
-TEXT,hyphen,TEXT
-TEXT,underscore,TEXT
-TEXT,space,TEXT
-TEXT,open_brace,OPEN_START_TAG_FROM_ATTRIBUTE
-OPEN_START_TAG_FROM_ATTRIBUTE,open_brace,OPEN_START_TAG_NAME
-OPEN_START_TAG_FROM_ATTRIBUTE,forward_slash,OPEN_END_TAG
+START,newline,START
+START,open_brace,OPEN_START_TAG
+START,close_brace,CLOSE_TAG
+START,letter,STRING
+START,digit,STRING
+START,hyphen,STRING
+START,underscore,STRING
+START,single_quotation_mark,STRING
+START,double_quotation_mark,STRING
+START,dot,STRING
+START,comma,STRING
+START,exclamation_mark,STRING
+START,equal,EQUAL
+
+OPEN_START_TAG,forward_slash,OPEN_CLOSE_TAG
+
+STRING,letter,STRING
+STRING,digit,STRING
+STRING,hyphen,STRING
+STRING,underscore,STRING
+STRING,single_quotation_mark,STRING
+STRING,double_quotation_mark,STRING
+STRING,dot,STRING
+STRING,comma,STRING
+
+
+
+
