--- src/tables/transition_table_new.csv
@@ -1,22 +1,0 @@
-Start_State,On,End_State
-START,space,START
-START,open_brace,OPEN_TAG
-START,close_brace,CLOSE_TAG
-START,letter,STRING
-START,digit,STRING
-START,hyphen,STRING
-START,underscore,STRING
-START,equal,EQUAL
-START,sinlge_quotation_mark,STRING
-START,double_quotation_mark,STRING
-STRING,space,END_STRING
-STRING,letter,STRING
-STRING,digit,STRING
-STRING,hyphen,STRING
-STRING,underscore,STRING
-STRING,equal,END_STRING
-STRING,sinlge_quotation_mark,STRING
-STRING,double_quotation_mark,STRING
-STRING,open_brace,END_STRING
-STRING,close_brace,END_STRING
-OPEN_TAG,forward_slash,OPEN_END_TAG
